Refactor: Extract new mod `utils/init.rs`
