+++ graphicsmagick/src/utils/init.rs
@@ -1,0 +1,34 @@
+use graphicsmagick_sys::InitializeMagick;
+use std::{ptr::null, sync::Once, thread};
+
+static HAS_INITIALIZED: Once = Once::new();
+
+/// Wrapper of `graphicsmagick_sys::InitializeMagick`, call it before any `graphicsmagick` action.
+/// Must be call in the main thread.
+pub fn initialize() {
+    HAS_INITIALIZED.call_once(|| {
+        assert_eq!(
+            thread::current().name(),
+            Some("main"),
+            "You have to call `graphicsmagick::initialize` in main thread"
+        );
+
+        unsafe {
+            InitializeMagick(null());
+        }
+    });
+}
+
+/// Check if [`initialize`] has called.
+#[inline]
+pub fn has_initialized() -> bool {
+    HAS_INITIALIZED.is_completed()
+}
+
+#[inline]
+pub(crate) fn assert_initialized() {
+    assert!(
+        has_initialized(),
+        "You have to call `graphicsmagick::initialize` first of all"
+    )
+}
