+++ graphicsmagick/src/utils/mod.rs
@@ -1,4 +1,3 @@
-use graphicsmagick_sys::InitializeMagick;
 use std::{
     borrow::Cow,
     ffi::{CStr, CString},
@@ -6,44 +5,14 @@ use std::{
     mem,
     ops::{Deref, DerefMut},
     os::raw::{c_char, c_double, c_uint, c_void},
-    ptr::{null, NonNull},
+    ptr::NonNull,
     slice::{from_raw_parts, from_raw_parts_mut},
     str::Utf8Error,
-    sync::Once,
-    thread,
 };
 
-static HAS_INITIALIZED: Once = Once::new();
-
-/// Wrapper of `graphicsmagick_sys::InitializeMagick`, call it before any `graphicsmagick` action.
-/// Must be call in the main thread.
-pub fn initialize() {
-    HAS_INITIALIZED.call_once(|| {
-        assert_eq!(
-            thread::current().name(),
-            Some("main"),
-            "You have to call `graphicsmagick::initialize` in main thread"
-        );
-
-        unsafe {
-            InitializeMagick(null());
-        }
-    });
-}
-
-/// Check if [`initialize`] has called.
-#[inline]
-pub fn has_initialized() -> bool {
-    HAS_INITIALIZED.is_completed()
-}
-
-#[inline]
-pub(crate) fn assert_initialized() {
-    assert!(
-        has_initialized(),
-        "You have to call `graphicsmagick::initialize` first of all"
-    )
-}
+mod init;
+pub(crate) use init::assert_initialized;
+pub use init::{has_initialized, initialize};
 
 /// For [`max_rgb`] to return max RGB value.
 pub trait MaxRGB {
